test : ServiceLayer 분리로 인한 TestCase 수정
