+++ src/test/java/bleuauction/bleuauction_be/server/member/service/MemberComponentServiceTest.java
@@ -1,7 +1,7 @@
 package bleuauction.bleuauction_be.server.member.service;
 
 
-import bleuauction.bleuauction_be.server.attach.service.AttachService;
+import bleuauction.bleuauction_be.server.attach.service.AttachComponentService;
 import bleuauction.bleuauction_be.server.common.jwt.CreateJwt;
 import bleuauction.bleuauction_be.server.common.jwt.TokenMember;
 import bleuauction.bleuauction_be.server.member.dto.LoginResponseDto;
@@ -40,7 +40,7 @@ class MemberComponentServiceTest {
     @Mock
     private MemberModuleService memberModuleService;
     @Mock
-    private AttachService attachService;
+    private AttachComponentService attachComponentService;
     @Mock
     private NcpObjectStorageService ncpObjectStorageService;
